requirements.txt
openai
python-dotenv
newspaper3k
lxml_html_clean